1. Lowered field update task from 5 seconds to 1. 2. Removed LastSpawn and added NextSpawn so mobs spawn 5 seconds after they die. With LastSpawn, a mob would spawn instantly as long as it is killed after 5 seconds of being spawned. 3. Added a class to calculate java's equivalent to System.currentTimeMilliseconds() as Time.GetCurrent() and Time.GetFuture(int seconds)

+++ Common/Wz/MobWz.cs
@@ -134,7 +134,7 @@ namespace NineToFive.Wz {
                             template.Level = ((WzIntProperty) property).Value;
                             break;
                         case "exp":
-                            template.Exp = (uint) ((WzIntProperty) property).Value;
+                            template.Exp = ((WzIntProperty) property).Value;
                             break;
                         case "hpRecovery":
                             template.HpRecovery = ((WzIntProperty) property).Value;
